spiel: Implement configuration database update using spiel API
* Implement configuration database update via spiel API.
* Add m0_confx_populate_all() that fills m0_confx structure from conf cache.
* Implement unit tests.

Configuration database update implementation consists of three parts:
* Construct in-memory configuration database using spiel API. (spiel/conf_mgmt.c)
* Send configuration to confd instances (LOAD phase). Corresponding FOPs and FOMs
  are implemented in conf/load_fop.c, conf/load_fom.c.
* Switch configuration on confd instances (FLIP phase). Corresponding FOPs and FOMs
  are implemented in conf/flip_fop.c, conf/flip_fom.c.
* Confd configure file with other version place as IO STOB

* Extract to separate function to get confd parameters
* Collect all QUOTE defines to M0_QUOTE
* Add m0_fop_domain_get and m0_fop_tm_get function
* Add m0_conf_version function

+++ conf/onwire.h
@@ -158,7 +158,7 @@ struct m0_confx_sdev {
 	/* Media type.  See m0_cfg_storage_device_media_type. */
 	uint32_t               xd_media;
 	/* Block size in bytes. */
-	uint64_t               xd_bsize;
+	uint32_t               xd_bsize;
 	/* Size in bytes. */
 	uint64_t               xd_size;
 	/* Last known state.  See m0_cfg_state_bit. */
